pyproject.toml
[tool.poetry]
name = "ppi-report-generator"
version = "0.1.0"
description = ""
authors = ["cnmorocho <carlos.nahuel.morocho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
openpyxl = "^3.1.5"
gspread = "^6.1.4"
google-api-python-client = "^2.158.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
